+++ Changes.md
@@ -4,7 +4,7 @@
 API
 ---
 
-- PlugAlgo : Added support in `setValueFromData()` for setting StringPlug values from StringVectorData.
+- PlugAlgo : Added support in `setValueFromData()` for setting StringPlug values from StringVectorData and StringVectorDataPlugValues from StringData.
 
 1.5.11.0 (relative to 1.5.10.1)
 ========